fix: responsive sizes
