+++ src/components/NavBar/index.tsx
@@ -1,172 +1,22 @@
 import {
-  Menu as MenuIcon,
-  PersonOutline as PersonOutlineIcon,
-  Settings as SettingsIcon
+  Menu as MenuIcon
 } from "@mui/icons-material";
 
 import {
   AppBar,
-  Box,
-  CssBaseline,
-  Drawer as MuiDrawer,
-  IconButton,
-  List, ListItem, ListItemButton, ListItemIcon, ListItemText, Menu, MenuItem, Toolbar
+  Box, Drawer as MuiDrawer,
+  IconButton, Toolbar
 } from "@mui/material";
 
 import {
-  navbarDrawerData,
-  UserRole
-} from "./navbarDrawerData";
-
-import {
-  UserAvatar,
-  UserDataCard,
-  UserDataList,
-  UserLabel, UserSettingsButton
-} from "./styles";
-
-import {
-  MouseEvent,
   useState
 } from "react";
 
 import { Outlet } from "react-router-dom";
-import { useNotification } from "../../context/notification";
+import { Drawer } from "./Drawer";
 
 
 const drawerWidth = 240;
-
-function Drawer() {
-  const [isUserModalOpen, setIsUserModalOpen] =
-    useState(false);
-  const toggleIsUserModalOpen = () =>
-    setIsUserModalOpen(!isUserModalOpen);
-
-  // TODO: Fetch User Data
-  const currentUser: {
-    role: UserRole,
-    name: string
-  } = {
-    role: "admin",
-    name: "User Name"
-  };
-
-  const { inform, notify } = useNotification();
-
-  const [userModalAnchorEl, setUserModalAnchorEl] =
-    useState<HTMLElement | null>(null);
-
-  const handleOpenUserModal = (event: MouseEvent<HTMLElement>) => {
-    setUserModalAnchorEl(event.currentTarget);
-  };
-
-  const handleCloseUserModal = () => {
-    setUserModalAnchorEl(null);
-  };
-
-  function provoke() {
-    try {
-      throw new Error("You asked for it...");
-    } catch (err: unknown) {
-      notify(err);
-    }
-  }
-
-  return (
-    <div>
-      <Box
-        style={{
-          marginTop: 50,
-          marginBottom: 30,
-          display: "flex",
-          justifyContent: "center",
-        }}
-      >
-        <img
-          className="logo"
-          src="/venus.svg"
-          style={{
-            height: 75
-          }}
-          alt="Logo"
-        />
-      </Box>
-      <List>
-
-        {navbarDrawerData()
-          .filter(({ userRoles }) => userRoles.includes(currentUser.role))
-          .map(({ navigate, title, Icon }, index) => (
-            <ListItem
-              key={index}
-              onClick={navigate}
-            >
-              <ListItemButton>
-                <ListItemIcon>
-                  <Icon />
-                </ListItemIcon>
-                <ListItemText primary={title} />
-              </ListItemButton>
-            </ListItem>
-          ))}
-
-        <UserDataList>
-          <UserDataCard>
-
-            <UserAvatar>
-              <PersonOutlineIcon />
-            </UserAvatar>
-
-            <UserLabel>
-              {currentUser.name}
-            </UserLabel>
-
-            <UserSettingsButton
-              size="large"
-              aria-label="account of current user"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenUserModal}
-              color="inherit"
-            >
-              <SettingsIcon />
-            </UserSettingsButton>
-
-            <Menu
-              id="menu-appbar"
-              anchorEl={userModalAnchorEl}
-              anchorOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              open={Boolean(userModalAnchorEl)}
-              onClose={handleCloseUserModal}
-            >
-              <MenuItem
-                onClick={() => inform("Not implemented")}
-              >
-                Profile
-              </MenuItem>
-              <MenuItem
-                onClick={provoke}
-              >
-                Exit
-              </MenuItem>
-
-            </Menu>
-
-          </UserDataCard>
-        </UserDataList>
-      </List>
-    </div>
-  );
-}
-
-
 
 
 export function NavBar() {
@@ -176,14 +26,14 @@ export function NavBar() {
 
   return (
     <Box sx={{ display: "flex" }}>
-      <CssBaseline />
+      {/* <CssBaseline /> */}
       <AppBar
         position="fixed"
         sx={{
           width: { sm: `calc(100% - ${drawerWidth}px)` },
           ml: { sm: `${drawerWidth}px` },
           backgroundColor: "transparent",
-          boxShadow: "none"
+          boxShadow: "none",
         }}
       >
         <Toolbar>
@@ -207,8 +57,12 @@ export function NavBar() {
 
       <Box
         component="nav"
-        sx={{ width: { sm: drawerWidth }, flexShrink: { sm: 0 } }}
-        aria-label="mailbox folders"
+        sx={{
+          width: {
+            sm: drawerWidth
+          },
+          flexShrink: { sm: 0 },
+        }}
       >
         <MuiDrawer
           variant="temporary"
@@ -222,7 +76,7 @@ export function NavBar() {
             "& .MuiDrawer-paper": {
               boxSizing: "border-box",
               width: drawerWidth
-            },
+            }
           }}
         >
           <Drawer />
@@ -244,17 +98,18 @@ export function NavBar() {
 
       </Box>
 
-      <Outlet />
 
       <Box
         component="main"
         sx={{
           flexGrow: 1,
-          height: "100vh",
-          overflow: "auto",
+          p: 3,
+          width: { sm: `calc(100% - ${drawerWidth}px)` },
+          // backgroundColor: "white"
         }}
-      />
-
+      >
+        <Outlet />
+      </Box>
     </Box>
   );
 }
