+++ src/main.tsx
@@ -2,6 +2,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { NotificationProvider } from "./context/notification";
+import { ThemeProvider } from "./context/theme";
 import "./index.css";
 import { RoutesProvider } from "./routes/RoutesProvider";
 import { ApiProvider } from "./services";
@@ -13,6 +14,7 @@ ReactDOM.createRoot(
   .render(
     [
       <React.StrictMode />,
+      <ThemeProvider />,
       <ApiProvider />,
       <NotificationProvider />,
       <RoutesProvider />
